[1433] - Fix dynamic front on settings page
Issue: https://github.com/mozilla-mobile/focus-ios/issues/1433

newLabel and accessoryLabel were not respecting the iOS Dynamic Font properties. This commit adds preferred font to both UILabels and fixes the issue.

This also adds a listener for when the font size changes, since adjustsFontForContentSizeCategory[1] only exists on iOS 10 and Focus has a deployment taget of iOS 9, an if #available was added to handle both situations

[1] https://developer.apple.com/documentation/uikit/uicontentsizecategoryadjusting/1771731-adjustsfontforcontentsizecategor?language=objc