Improves contrast of action sheets when accessibility features enabled.
The Increase Contrast and Reduce Transparency accessibility features
should improve accessibility. However, before this change enabling them
caused the background of action sheets to become medium grey and the
text to become a slightly lighter grey, which made the text hard to
read. This prevents the problem by avoiding the use of a blur effect
view when Reduce Transparency is enabled and using white text when
Increase Contrast is enabled to prevent it from being darkened.