Fixes #2004 - Custom autocompletions have a / suffix
Domain completions currently append `/` to all matched domains. This leads to unexpected
behavior when the completion isn't just a domain, but contains a path as well. This patch
changes the behavior to append a `/` only if no `/` is present in the matchedDomain string.
If such a character is present, we can safely leave the string unaltered.

Closes #2004.